+++ src/pages/GrupoAMIE/GrupoAMIE.js
@@ -3,8 +3,6 @@ import { Button, View, Text, Alert } from "react-native";
 import YoutubePlayer from "react-native-youtube-iframe";
 import {WebView} from "react-native-webview";
 import {
-  Text,
-  View,
   TouchableOpacity,
   useWindowDimensions,
   ScrollView,
@@ -24,18 +22,6 @@ import {
 } from "./Styles";
 
 function GrupoAMIE() {
-  const [playing, setPlaying] = useState(false);
-
-  const onStateChange = useCallback((state) => {
-    if (state === "ended") {
-      setPlaying(false);
-      Alert.alert("video has finished playing!");
-    }
-  }, []);
-
-  const togglePlaying = useCallback(() => {
-    setPlaying((prev) => !prev);
-  }, []);
   
   const { width } = useWindowDimensions();
   const { height } = useWindowDimensions();
@@ -80,16 +66,6 @@ function GrupoAMIE() {
         </Body>
       </ScrollView>
     </Container>
-    <View>
-      <Text>GrupoAMIE</Text>
-    <YoutubePlayer
-    javaScriptEnabled={true}
-    height={300}
-    play={playing}
-    videoId={"LOrALdmPWVU"}
-    onChangeState={onStateChange}
-  />
-  </View>
   );
 }
